requirements.txt
streamlit==1.28.0
langchain==0.1.0
langchain-google-genai
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain-community
redis